Change analysis parameters

Change analysis parameters.

changeParameter(x, parameterName, elements = analysisElements()) <- value

# S4 method for AnalysisParameters
changeParameter(x, parameterName, elements = analysisElements()) <- value

Arguments

x

S4 object of class AnalysisParameters

parameterName

name of the parameter to change

elements

character vector of analysis elements to target parameter change. Can be any returned by analysisElements().

value

New value of the parameter

Value

An S4 object of class AnalysisParameters.

Details

For the parameter name selected, all parameters with that name will be altered.

Examples

p <- analysisParameters('pre-treatment')

changeParameter(p,'cls') <- 'day'

print(p)
#> Parameters:
#> pre-treatment
#> 	QC
#> 		occupancyFilter
#> 			cls = day
#> 			QCidx = QC
#> 			occupancy = 2/3
#> 		impute
#> 			cls = day
#> 			QCidx = QC
#> 			occupancy = 2/3
#> 			parallel = variables
#> 			seed = 1234
#> 		RSDfilter
#> 			cls = day
#> 			QCidx = QC
#> 			RSDthresh = 50
#> 		removeQC
#> 			cls = day
#> 			QCidx = QC
#> 	occupancyFilter
#> 		maximum
#> 			cls = day
#> 			occupancy = 2/3
#> 	impute
#> 		class
#> 			cls = day
#> 			occupancy = 2/3
#> 			seed = 1234
#> 	transform
#> 		TICnorm
#>