Feature correlation analysis

Feature correlation analysis.

correlations(d, ...)

# S4 method for AnalysisData
correlations(
  d,
  method = "pearson",
  pAdjustMethod = "bonferroni",
  corPvalue = 0.05
)

# S4 method for Analysis
correlations(d)

Arguments

d
S4 object of class AnalysisData

...
arguments to pass to specific method

method
correlation method. One of pearson or spearman.

pAdjustMethod
p-value adjustment method. See ?p.adjust for available methods.

corPvalue
p-value cut-off threshold for significance

Value

A tibble containing results of significantly correlated features.

Details

Correlation analyses can be used to identify associated features within data sets. This can be useful to identifying clusters of related features that can be used to annotate metabolites within data sets. All features are compared and the returned table of correlations are p-value thresholded using the specified cut-off.

Examples

library(metaboData)

d <- analysisData(abr1$neg[,200:300],abr1$fact)

correlations(d)

#> # A tibble: 130 x 6
#>    Feature1 Feature2 log2IntensityRatio     r        p     n
#>    <chr>    <chr>                 <dbl> <dbl>    <dbl> <int>
#>  1 N213     N214                 1.69   0.591 3.40e- 5    63
#>  2 N213     N218                 4.02   0.741 1.21e- 2    21
#>  3 N209     N224                 1.06   0.810 6.71e- 7    34
#>  4 N213     N224                 2.57   0.807 2.75e- 7    36
#>  5 N213     N225                 1.78   0.610 1.15e- 7    82
#>  6 N214     N225                 0.0941 0.477 2.30e- 2    55
#>  7 N224     N225                -0.792  0.863 1.76e- 7    29
#>  8 N212     N227                -0.884  0.980 1.07e- 2     7
#>  9 N213     N229                -2.30   0.682 6.66e-14   107
#> 10 N214     N229                -3.99   0.456 1.07e- 2    67
#> # … with 120 more rows