plotRSD

Plot RSD distributions of raw data in quality control samples.

plotRSD(analysis, cls = "class", ...)

# S4 method for AnalysisData
plotRSD(analysis, cls = "class")

# S4 method for Analysis
plotRSD(
  analysis,
  cls = "class",
  QCidx = "QC",
  QCparameters = NULL,
  histBins = 30,
  title = ""
)

Arguments

analysis
object of class Analysis containing analysis results

cls
info column to use for class labels

...
arguments to pass to the appropriate method

QCidx
QC sample label

QCparameters
alternative parameters for QC sample pre-treatment. See details

histBins
number of bins to use for histogram plotting

title
plot title

Details

If QCparameters is set as NULL, the default QC pre-treatment parameters are used as given by analysisParameters('preTreat'). Alternative pre-treatment routines can be used by specifying an AnalysisParameters object for QCparameters.

Examples

if (FALSE) {

library(metaboData)
library(binneR)

files <- filePaths("FIE-HRMS","BdistachyonEcotypes")

info <- runinfo("FIE-HRMS","BdistachyonEcotypes")

binDat <- binneRlyse(files, 
                       info, 
                       parameters = detectParameters(files))

p <- analysisParameters('preTreat')

analysis <- metabolyse(binnedData(binDat)$n,info(binDat),p)

plotRSD(analysis)
}