reAnalyse

Re-analyse an object of class Analysis using specified parameters.

reAnalyse(analysis, parameters = analysisParameters(), verbose = TRUE)

Arguments

analysis
an object of class Analysis containing previous analysis results

parameters
an object of class Parameters containing parameters for re-analysis

verbose
should output be printed to the console

Examples

library(metaboData)

## Generate analysis parameters
p <- analysisParameters(c('pre-treatment','modelling'))

## Alter pre-treatment and modelling parameters to use different methods
parameters(p,'pre-treatment') <- preTreatmentParameters(
  list(occupancyFilter = 'maximum',
       transform = 'TICnorm')
)
parameters(p,'modelling') <- modellingParameters('anova')

## Change "cls" and "nCores" parameters 
changeParameter(p,'cls') <- 'day'
changeParameter(p,'nCores') <- 2

## Run analysis using a subset of the abr1 negative mode data set
analysis <- metabolyse(abr1$neg[,1:200],
                       abr1$fact,
                       p)

#> 
#> metabolyseR  v0.13.2 Mon Nov  2 15:51:12 2020

#> ________________________________________________

#> Parameters:
#> pre-treatment
#> 	occupancyFilter
#> 		maximum
#> 			cls = day
#> 			occupancy = 2/3
#> 	transform
#> 		TICnorm
#> 
#> modelling
#> 	anova
#> 		cls = day
#> 		pAdjust = bonferroni
#> 		comparisons = list()
#> 		returnModels = FALSE
#> 		nCores = 2
#> 		clusterType = getClusterType()

#> ________________________________________________

#> Pre-treatment …

#> Pre-treatment 	✓ [1.1S]

#> Modelling …

#> Modelling 	✓ [0.9S]

#> ________________________________________________

#> 
#> Complete! [2.2S]

                       
## Re-analyse to include correlation analysis
analysis <- reAnalyse(analysis,
                      parameters = analysisParameters('correlations'))

#> 
#> metabolyseR v0.13.2 Mon Nov  2 15:51:14 2020
#> ________________________________________________
#> Parameters:
#> correlations
#> 	method = pearson
#> 	pAdjustMethod = bonferroni
#> 	corPvalue = 0.05
#> ________________________________________________
#> 

#> Correlations …

#> Correlations 	✓ [0.2S]

#> ________________________________________________
#> 
#> Complete! [0.2S]
#>