Modelling accessor methods

Methods for accessing modelling results.

binaryComparisons(x, cls = "class")

# S4 method for AnalysisData
binaryComparisons(x, cls = "class")

type(x)

# S4 method for RandomForest
type(x)

response(x)

# S4 method for RandomForest
response(x)

metrics(x)

# S4 method for RandomForest
metrics(x)

# S4 method for list
metrics(x)

importanceMetrics(x)

# S4 method for RandomForest
importanceMetrics(x)

importance(x)

# S4 method for RandomForest
importance(x)

# S4 method for Univariate
importance(x)

# S4 method for list
importance(x)

# S4 method for Analysis
importance(x)

explanatoryFeatures(x, ...)

# S4 method for Univariate
explanatoryFeatures(x, threshold = 0.05, ...)

# S4 method for RandomForest
explanatoryFeatures(x, metric = "FalsePositiveRate", threshold = 0.05)

# S4 method for list
explanatoryFeatures(x, threshold = 0.05, ...)

# S4 method for Analysis
explanatoryFeatures(x, threshold = 0.05, ...)

Arguments

x
S4 object of class AnalysisData,RandomForest, Univariate, Analysis or a list.

cls
sample information column to use

...
arguments to parse to method for specific class

threshold
threshold below which explanatory features are extracted

metric
importance metric for which to retrieve explanatory features

Methods

binaryComparisons: Return a vector of all possible binary comparisons for a given sample information column.

type: Return the type of random forest analysis.

response: Return the response variable name used for a random forest analysis.

metrics: Retrieve the model performance metrics for a random forest analysis

importanceMetrics: Retrieve the available feature importance metrics for a random forest analysis.

importance: Retrieve feature importance results.

explanatoryFeatures: Retrieve explanatory features.

Examples

library(metaboData)

d <- analysisData(abr1$neg[,200:300],abr1$fact)

## Return possible binary comparisons for the 'day' column
binaryComparisons(d,cls = 'day')

#>  [1] "1~2" "1~3" "1~4" "1~5" "1~H" "2~3" "2~4" "2~5" "2~H" "3~4" "3~5" "3~H"
#> [13] "4~5" "4~H" "5~H"

## Perform random forest analysis
rf_analysis <- randomForest(d,cls = 'day')

## Return the type of random forest
type(rf_analysis)

#> [1] "classification"

## Return the response variable name used
response(rf_analysis)

#> [1] "day"

## Retrieve the model performance metrics
metrics(rf_analysis)

#> # A tibble: 4 x 5
#>   Response Comparison  .metric  .estimator .estimate
#>   <chr>    <chr>       <chr>    <chr>          <dbl>
#> 1 day      1~2~3~4~5~H accuracy multiclass    0.567 
#> 2 day      1~2~3~4~5~H kap      multiclass    0.48  
#> 3 day      1~2~3~4~5~H roc_auc  hand_till     0.886 
#> 4 day      1~2~3~4~5~H margin   NA            0.0424

## Show the available feature importance metrics
importanceMetrics(rf_analysis)

#>  [1] "1"                    "2"                    "3"                   
#>  [4] "4"                    "5"                    "FalsePositiveRate"   
#>  [7] "H"                    "MeanDecreaseAccuracy" "MeanDecreaseGini"    
#> [10] "SelectionFrequency"  

## Retrieve the feature importance results
importance(rf_analysis)

#> # A tibble: 1,010 x 5
#>    Response Comparison  Feature Metric                  Value
#>    <chr>    <chr>       <chr>   <chr>                   <dbl>
#>  1 day      1~2~3~4~5~H N200    1                    0       
#>  2 day      1~2~3~4~5~H N200    2                    0       
#>  3 day      1~2~3~4~5~H N200    3                    0       
#>  4 day      1~2~3~4~5~H N200    4                    0       
#>  5 day      1~2~3~4~5~H N200    5                    0       
#>  6 day      1~2~3~4~5~H N200    FalsePositiveRate    2.35e-40
#>  7 day      1~2~3~4~5~H N200    H                    0       
#>  8 day      1~2~3~4~5~H N200    MeanDecreaseAccuracy 0       
#>  9 day      1~2~3~4~5~H N200    MeanDecreaseGini     6.00e- 2
#> 10 day      1~2~3~4~5~H N200    SelectionFrequency   1.6 e+ 1
#> # … with 1,000 more rows

## Retrieve the explanatory features
explanatoryFeatures(rf_analysis,metric = 'FalsePositiveRate',threshold = 0.05)

#> # A tibble: 35 x 5
#>    Response Comparison  Feature Metric                Value
#>    <chr>    <chr>       <chr>   <chr>                 <dbl>
#>  1 day      1~2~3~4~5~H N229    FalsePositiveRate 5.75e-129
#>  2 day      1~2~3~4~5~H N259    FalsePositiveRate 4.88e- 72
#>  3 day      1~2~3~4~5~H N277    FalsePositiveRate 3.98e- 67
#>  4 day      1~2~3~4~5~H N255    FalsePositiveRate 3.27e- 53
#>  5 day      1~2~3~4~5~H N213    FalsePositiveRate 4.92e- 45
#>  6 day      1~2~3~4~5~H N200    FalsePositiveRate 2.35e- 40
#>  7 day      1~2~3~4~5~H N221    FalsePositiveRate 1.80e- 38
#>  8 day      1~2~3~4~5~H N299    FalsePositiveRate 4.91e- 36
#>  9 day      1~2~3~4~5~H N245    FalsePositiveRate 9.75e- 27
#> 10 day      1~2~3~4~5~H N279    FalsePositiveRate 2.38e- 20
#> # … with 25 more rows